+++ src/components/Card/Rating/index.test.tsx
@@ -1,4 +1,3 @@
-import '@testing-library/jest-dom';
 import { Rating } from '.';
 import { render, screen } from '@testing-library/react';
 
@@ -6,12 +5,9 @@ describe('<Rating /> test', () => {
   test('Should render correctly with granted data', () => {
     const value = 4.3;
 
-    const { container } = render(<Rating value={value} />);
-    const el = container.querySelector('.product-rating');
-    const stars = container.querySelectorAll('.product-rating__star .product-rating__star-inner');
+    render(<Rating value={value} />);
 
-    expect(el).not.toBeNull();
-    expect(stars.length).toEqual(Rating.NumOfStars);
+    expect(screen.getAllByRole('star').length).toEqual(Rating.NumOfStars);
     expect(screen.getByTitle(`rating ${value}`)).toBeInTheDocument();
   });
 });
