+++ src/components/SearchBar/style.scss
